Added middleware and fixed typing on user controller

+++ backend/src/controller/UserController.ts
@@ -1,6 +1,6 @@
-import { NextFunction, Request, Response } from 'express';
-import { User } from '../entity/User';
-import { AppDataSource } from '../data-source';
+import { NextFunction, Request, Response } from "express";
+import { User } from "../entity/User";
+import { AppDataSource } from "../data-source";
 
 export class UserController {
   private userRepository = AppDataSource.getRepository(User);
@@ -10,7 +10,11 @@ export class UserController {
   }
 
   async one(request: Request, response: Response, next: NextFunction) {
-    return this.userRepository.findOne(request.params.id);
+    return this.userRepository.findOne({
+      where: {
+        id: parseInt(request.params.id),
+      },
+    });
   }
 
   async save(request: Request, response: Response, next: NextFunction) {
@@ -18,9 +22,10 @@ export class UserController {
   }
 
   async remove(request: Request, response: Response, next: NextFunction) {
-    let userToRemove = await this.userRepository.findOneBy({
-      id: request.params.id,
+    const userToRemove = await this.userRepository.findOneBy({
+      id: parseInt(request.params.id),
     });
+    if (userToRemove) throw new Error("This User was not found");
     await this.userRepository.remove(userToRemove);
   }
 }
